Hacked apart delegate structure and state with specifiers - needs extensive testing and changes to test client if working
