+++ penne/delegates.py
@@ -1,45 +1,138 @@
-"""Module Defining Default Delegates and Delegate Related Classes"""
+"""Collection of Noodles Objects
+
+Follows the specification in the cddl document, and
+implements strict validation
+"""
 
 from __future__ import annotations
-from typing import TYPE_CHECKING, Callable
+from typing import Literal, Optional, Any, Union, Callable, TYPE_CHECKING
 if TYPE_CHECKING:
-    from penne.messages import Message
     from penne.core import Client
 
+from collections import namedtuple
+from enum import Enum
+from math import pi
 
-class Delegate(object):
+from pydantic import BaseModel, root_validator
+
+
+""" =============================== ID's ============================= """
+
+IDGroup = namedtuple("IDGroup", ["slot", "gen"])
+
+class ID(IDGroup):
+
+    __slots__ = ()
+    def __repr__(self):
+        return f"{self.__class__}|{self.slot}/{self.gen}|"
+
+    def __key(self):
+        return (type(self), self.slot, self.gen)
+
+    def __eq__(self, __o: object) -> bool:
+        if isinstance(__o, ID):
+            return self.__key() == __o.__key()
+        else:
+            return False
+
+    def __hash__(self):
+        return hash(self.__key())
+
+class MethodID(ID):
+    pass
+
+class SignalID(ID):
+    pass
+
+class EntityID(ID):
+    pass
+
+class PlotID(ID):
+    pass
+
+class BufferID(ID):
+    pass
+
+class BufferViewID(ID):
+    pass
+
+class MaterialID(ID):
+    pass
+
+class ImageID(ID):
+    pass
+
+class TextureID(ID):
+    pass
+
+class SamplerID(ID):
+    pass
+
+class LightID(ID):
+    pass
+
+class GeometryID(ID):
+    pass
+
+class TableID(ID):
+    pass
+
+
+""" ====================== Generic Parent Class ====================== """
+
+class NoodleObject(BaseModel):
+    """Parent Class for all noodle objects"""
+
+    class Config:
+        """Configuration for Validation"""
+
+        arbitrary_types_allowed = True
+        use_enum_values = True
+
+    def __repr__(self) -> str:
+        return f"{type(self)}"
+
+class Component(NoodleObject):
+    """Parent class for all components"""
+
+    id: ID = None
+
+    def __repr__(self):
+        return f"{type(self)} | {self.id}"
+
+class Delegate(NoodleObject):
     """Parent class for all delegates
     
     Defines general methods that should be available for all delegates
     
     Attributes:
         client (Client): Client delegate is attached to
-        info (Message): Message containing all info on delegate
-        specifier (str): Keyword specifying delegate in state
         signals (dict): Signals that can be called on delegate
         __all__ (list): Specify public methods, used in show_methods()
     """
 
-    def __init__(self, client: Client, message: Message, specifier: str):
-        self.client = client
-        self.info = message
-        self.specifier = specifier
-        self.signals = {}
-        self.__all__ = []
-    
+    client: Client
+    id: ID = None
+    name: Optional[str]
+    signals: dict = {}
+    __all__: list = []
+
+    def __repr__(self):
+        return f"{type(self)} | {self.id}"
+
     def __repr__(self) -> str:
-        return f"{self.specifier} delegate | {self.info.id}"
+        return f"{self.specifier} delegate | {self.id}"
 
     # For all except Document Delegate
-    def on_new(self, message: Message):
+    def on_new(self, message: dict):
         pass
 
     # For Document, Table, Entity, Plot, Material, Light Delegates
-    def on_update(self, message: Message):
+    def on_update(self, message: dict):
         pass
 
     # For all except Document Delegate
-    def on_remove(self, message: Message):
+    def on_remove(self, message: dict):
         pass
 
 
@@ -53,125 +146,200 @@ class Delegate(object):
             print(f">> '{name}'\n{method.__doc__}")
 
 
-class InjectedMethod(object):
-    """Class for representing injected method in delegate
+""" ====================== Common Definitions ====================== """
 
-    Attributes:
-        method (method): method to be called
-        injected (bool): attribute marking method as injected
-    """
+Vec3 = tuple[float, float, float]
+Vec4 = tuple[float, float, float, float]
+Mat3 = tuple[float, float, float, 
+             float, float, float, 
+             float, float, float]
+Mat4 = tuple[float, float, float, float,
+             float, float, float, float,
+             float, float, float, float,
+             float, float, float, float]
 
-    def __init__(self, method_obj) -> None:
-        self.method = method_obj
-        self.injected = True
+RGB = Vec3
+RGBA = Vec4
 
-    def __call__(self, *args, **kwds):
-        self.method(*args, **kwds)
+class AttributeSemantic(Enum):
+    position = "POSITION"
+    normal = "NORMAL"
+    tangent = "TANGENT"
+    texture = "TEXTURE"
+    color = "COLOR"
 
+class Format(Enum):
+    u8 = "U8"
+    u16 = "U16"
+    u32 = "U32"
+    u8vec4 = "U8VEC4"
+    u16vec2 = "U16VEC2"
+    vec2 = "VEC2"
+    vec3 = "VEC3"
+    vec4 = "VEC4"
+    mat3 = "MAT3"
+    mat4 = "MAT4"
 
-class LinkedMethod(object):
-    """Class linking target delegate and method's delegate 
+class PrimitiveType(Enum):
+    points = "POINTS"
+    lines = "LINES"
+    line_loop = "LINE_LOOP"
+    line_strip = "LINE_STRIP"
+    triangles = "TRIANGLES"
+    triangle_strip = "TRIANGLE_STRIP"
+
+class SamplerMode(Enum):
+    clamp_to_edge = "CLAMP_TO_EDGE"
+    mirrored_repeat = "MIRRORED_REPEAT"
+    repeat = "REPEAT"
+
+class URL(NoodleObject):
+    url: str
+
+class SelectionRange(NoodleObject):
+    key_from_inclusive: int
+    key_to_exclusive: int
+
+class Selection(NoodleObject):
+    name: str
+    rows: Optional[list[int]] = None
+    row_ranges: Optional[list[SelectionRange]] = None
+
+class MethodArg(NoodleObject): 
+    name: str
+    doc: Optional[str] = None 
+    editor_hint: Optional[str] = None
+
+class BoundingBox(NoodleObject):
+    min: Vec3
+    max: Vec3
+
+class TextRepresentation(NoodleObject):
+    txt: str
+    font: Optional[str] = "Arial"
+    height: Optional[float] = .25
+    width: Optional[float] = -1.0
+
+class WebRepresentation(NoodleObject):
+    source: str
+    height: Optional[float] = .5
+    width: Optional[float] = .5
+
+class InstanceSource(NoodleObject):
+    view: BufferViewID # view of mat4
+    stride: int 
+    bb: Optional[BoundingBox] = None
+
+class RenderRepresentation(NoodleObject):
+    mesh: GeometryID
+    instances: Optional[InstanceSource] = None
+
+class TextureRef(NoodleObject):
+    texture: TextureID
+    transform: Optional[Mat3] = [1.0, 0.0, 0.0,
+                       0.0, 1.0, 0.0,
+                       0.0, 0.0, 1.0,]
+    texture_coord_slot: Optional[int] = 0.0
+
+class PBRInfo(NoodleObject):
+    base_color: RGBA = [1.0, 1.0, 1.0, 1.0]
+    base_color_texture: Optional[TextureRef] = None # assume SRGB, no premult alpha
+
+    metallic: Optional[float] = 1.0
+    roughness: Optional[float] = 1.0
+    metal_rough_texture: Optional[TextureRef] = None # assume linear, ONLY RG used
+
+class PointLight(NoodleObject):
+    range: float = -1.0
+
+class SpotLight(NoodleObject):
+    range: float = -1.0
+    inner_cone_angle_rad: float = 0.0
+    outer_cone_angle_rad: float = pi/4
+
+class DirectionalLight(NoodleObject):
+    range: float = -1.0
+
+class Attribute(NoodleObject):
+    view: BufferViewID
+    semantic: AttributeSemantic
+    channel: Optional[int] = None
+    offset: Optional[int] = 0
+    stride: Optional[int] = 0
+    format: Format
+    minimum_value: Optional[list[float]] = None
+    maximum_value: Optional[list[float]] = None
+    normalized: Optional[bool] = False
+
+class Index(NoodleObject):
+    view: BufferViewID 
+    count: int
+    offset: Optional[int] = 0
+    stride: Optional[int] = 0
+    format: Literal["U8", "U16", "U32"]
+
+class GeometryPatch(NoodleObject):
+    attributes: list[Attribute]
+    vertex_count: int
+    indices: Optional[Index] = None
+    type: PrimitiveType
+    material: MaterialID # Material ID
+
+class InvokeIDType(NoodleObject):
+    entity: Optional[EntityID] = None
+    table: Optional[TableID] = None
+    plot: Optional[PlotID] = None
+
+    @root_validator
+    def one_of_three(cls, values):
+        already_found  = False
+        for field in values:
+            if values[field] and already_found:
+                raise ValueError("More than one field entered")
+            elif values[field]:
+                already_found = True
         
-    make a cleaner function call in injected method
-    
-    Attributes:
-        _obj_delegate (delegate): 
-            delgate method is being linked to
-        _method_delegate (MethodDelegate): 
-            the method's delegate 
-    """
+        if not already_found:
+            raise ValueError("No field provided")
+        else:
+            return values
 
-    def __init__(self, object_delegate: Delegate, method_delegate: Delegate):
-        self._obj_delegate = object_delegate
-        self._method_delegate = method_delegate
+class TableColumnInfo(NoodleObject):
+    name: str
+    type: Literal["TEXT", "REAL", "INTEGER"]
 
-    def __call__(self, on_done=None, *arguments):
-        self._method_delegate.invoke(self._obj_delegate, arguments, callback=on_done)
+class TableInitData(NoodleObject):
+    columns: list[TableColumnInfo]
+    keys: list[int]
+    data: list[list[Union[float, int, str]]]
+    selections: Optional[list[Selection]] = None
 
-
-def inject_methods(delegate: Delegate, methods: list):
-    """Inject methods into a delegate class
-
-    Args:
-        delegate_name (str): 
-            identifier for delegate to be modified
-        methods (list): 
-            list of method id's to inject
-    """
-
-    # Clear out old injected methods
-    for name in dir(delegate):
-        att = getattr(delegate, name)
-        if hasattr(att, "injected"):
-            print(f"Deleting: {name} in inject methods")
-            delattr(delegate, name)
-
-    state_methods = delegate.client.state["methods"] 
-    for id in methods:
-
-        # Get method delegate and manipulate name to exclude noo::
-        method = state_methods[tuple(id)]
-        name = method.info.name[5:]
-
-        # Create injected by linking delegates, and creating call method
-        linked = LinkedMethod(delegate, method)
-        injected = InjectedMethod(linked.__call__)
-
-        setattr(delegate, name, injected)
+    # too much overhead? - strict mode
+    @root_validator
+    def types_match(cls, values):
+        for row in values['data']:
+            for col, i in zip(values['columns'], range(len(row))):
+                text_mismatch = isinstance(row[i], str) and col.type != "TEXT"
+                real_mismatch = isinstance(row[i], float) and col.type != "REAL"
+                int_mismatch = isinstance(row[i], int) and col.type != "INTEGER"
+                if text_mismatch or real_mismatch or int_mismatch:
+                    raise ValueError(f"Column Info doesn't match type in data: {col, row[i]}")
+        return values
 
 
-def inject_signals(delegate: Delegate, signals: list[list[int]]):
-    """Method to inject signals into delegate
+class Document(NoodleObject):
+    pass
+        
 
-    Args:
-        delegate (delegate): 
-            delegate object to be injected 
-        signals (list): 
-            list of signal id's to be injected
-    """
-
-    state_signals = delegate.client.state["signals"]
-    injected_signals = {}
-    for id in signals:
-        signal: SignalDelegate = state_signals[tuple(id)]
-        injected_signals[signal.info.name] = None
-    delegate.signals = injected_signals
+""" ====================== NOOODLE COMPONENTS ====================== """
 
 
-
-"""
-Default Delegates for Python Client
-"""
-
-class MethodDelegate(Delegate):
-    """Delegate representing a method which can be invoked on the server
-
-    Attributes:
-        client (client object): 
-            client delegate is a part of 
-        info (message): 
-            message containing information on the method
-        specifier (str): 
-            keyword for specifying the type of delegate
-        context_map (dict):
-            mapping specifier to context for method invocation
-    """
-
-    def __init__(self, client: Client, message: Message, specifier: str):
-        self.client = client
-        self.info = message
-        self.specifier = specifier
-        self.context_map = {
-            "tables": "table",
-            "plots": "plot",
-            "entities": "entity"
-        }
-
-    def on_new(self, message: Message):
-        pass
-
-    def on_remove(self, message: Message):
-        pass
+class Method(Delegate):
+    id: MethodID
+    name: str
+    doc: Optional[str] = None
+    return_doc: Optional[str] = None
+    arg_doc: list[MethodArg] = []
 
     def invoke(self, on_delegate: Delegate, args=None, callback=None):
         """Invoke this delegate's method
@@ -185,118 +353,211 @@ class MethodDelegate(Delegate):
             callback (function):
                 function to be called when complete
         """
-        context = {self.context_map[on_delegate.specifier]: on_delegate.info.id}
-        self.client.invoke_method(self.info.id, args, context=context, on_done=callback)
+        context_map = {
+            Table: "table",
+            Plot: "plot",
+            Entity: "entity"
+        }
+        context = {context_map[type(on_delegate)]: on_delegate.id}
+        self.client.invoke_method(self.id, args, context=context, on_done=callback)
 
 
     def __repr__(self) -> str:
         """Custom string representation for methods"""
         
-        rep = f"{self.info.name}:\n\t{self.info.doc}\n\tReturns: {self.info.return_doc}\n\tArgs:"
-        for arg in self.info.arg_doc:
+        rep = f"{self.name}:\n\t{self.doc}\n\tReturns: {self.return_doc}\n\tArgs:"
+        for arg in self.arg_doc:
             rep += f"\n\t\t{arg.name}: {arg.doc}"
         return rep
 
 
-class SignalDelegate(Delegate):
-    """Delegate representing a signal coming from the server
+class Signal(Delegate):
+    id: SignalID
+    name: str
+    doc: Optional[str] = None
+    arg_doc: list[MethodArg] = None
 
-    Attributes:
-        client (Client): 
-            client delegate is a part of 
-        info (message): 
-            message containing information on the signal
-        specifier (str): 
-            keyword for specifying the type of delegate
-    """
+
+class Entity(Delegate):
+    id: EntityID
+    name: Optional[str] = "Unnamed Entity Delegate"
+
+    parent: Optional[EntityID] = None
+    transform: Optional[Mat4] = None
+
+    text_rep: Optional[TextRepresentation] = None
+    web_rep: Optional[WebRepresentation] = None
+    render_rep: Optional[RenderRepresentation] = None
+
+    lights: Optional[list[LightID]] = None
+    tables: Optional[list[TableID]] = None
+    plots: Optional[list[PlotID]] = None
+    tags: Optional[list[str]] = None
+    methods_list: Optional[list[MethodID]] = None
+    signals_list: Optional[list[SignalID]] = None
+
+    influence: Optional[BoundingBox] = None
+
+
+class Plot(Delegate):
+    id: PlotID
+    name: Optional[str] = "Unnamed Plot Delegate"
+
+    table: Optional[TableID] = None
+
+    simple_plot: Optional[str] = None
+    url_plot: Optional[str] = None
+
+    methods_list: Optional[list[MethodID]] = None
+    signals_list: Optional[list[SignalID]] = None
+
+    @root_validator
+    def one_of(cls, values):
+        if bool(values['simple_plot']) != bool(values['url_plot']):
+            return values
+        else:
+            raise ValueError("One plot type must be specified")
+
+
+class Buffer(Delegate):
+    id: BufferID
+    name: Optional[str] = "Unnamed Buffer Delegate"
+    size: int = None
+
+    inline_bytes: bytes = None
+    uri_bytes: str = None
+
+    @root_validator
+    def one_of(cls, values):
+        if bool(values['inline_bytes']) != bool(values['uri_bytes']):
+            return values
+        else:
+            raise ValueError("One plot type must be specified")
+
+
+class BufferView(Delegate):
+    id: BufferViewID
+    name: Optional[str] = "Unnamed Buffer-View Delegate"    
+    source_buffer: BufferID
+
+    type: Literal["UNK", "GEOMETRY", "IMAGE"]
+    offset: int
+    length: int
+
     
-    def __init__(self, client: Client, message: Message, specifier: str):
-        self.client = client
-        self.info = message
-        self.specifier = specifier
+class Material(Delegate):
+    id: MaterialID
+    name: Optional[str] = "Unnamed Material Delegate"
 
-    def on_new(self, message: Message):
-        pass
+    pbr_info: Optional[PBRInfo] = PBRInfo()
+    normal_texture: Optional[TextureRef] = None
 
-    def on_remove(self, message: Message): 
-        pass
+    occlusion_texture: Optional[TextureRef] = None # assumed to be linear, ONLY R used
+    occlusion_texture_factor: Optional[float] = 1.0
 
+    emissive_texture: Optional[TextureRef] = None # assumed to be SRGB, ignore A
+    emissive_factor: Optional[Vec3] = [1.0, 1.0, 1.0]
 
-class SelectionRange(tuple):
-    """Selection of range of rows"""
+    use_alpha: Optional[bool] = False
+    alpha_cutoff: Optional[float] = .5
 
-    def __new__(cls, key_from: int, key_to: int):
-        return super().__new__(SelectionRange, (key_from, key_to))
-
-
-class Selection(object):
-    """Selection of certain rows in a table
-
-    Attributes:
-        name (str): 
-            name of the selection
-        rows (list[int]): 
-            list of indices of rows
-        row_ranges (list[SelectionRange]): 
-            ranges of selected rows
-    """
-
-    def __init__(self, name: str, rows: list[int] = None, row_ranges: list[SelectionRange] = None):
-        self.name = name
-        self.rows = rows
-        self.row_ranges = row_ranges
-
-    def __repr__(self) -> str:
-        return f"Selection Object({self.__dict__})"
-
-    def __getitem__(self, attribute):
-        return getattr(self, attribute)
+    double_sided: Optional[bool] = False
 
 
-class TableDelegate(Delegate):
-    """Delegate representing a table
+class Image(Delegate):
+    id: ImageID
+    name: Optional[str] = "Unnamed Image Delegate"
 
-    Each table delegate corresponds with a table on the server
-    To use the table, you must first subscribe 
+    buffer_source: BufferID = None
+    uri_source: str = None
 
-    Attributes:
-        client (Client): 
-            weak ref to client to invoke methods and such
-        dataframe (Dataframe): 
-            dataframe representing current state of the table
-        selections (dict): 
-            mapping of name to selection object
-        signals (signals): 
-            mapping of signal name to function
-        name (str): 
-            name of the table
-        id (list): 
-            id group for delegate in state and table on server
-    """
+    @root_validator
+    def one_of(cls, values):
+        if bool(values['buffer_source']) != bool(values['uri_source']):
+            return values
+        else:
+            raise ValueError("One plot type must be specified")
 
-    def __init__(self, client: Client, message: Message, specifier: str):
-        super().__init__(client, message, specifier)
-        self.name = "Table Delegate"
-        self.selections = {}
-        self.signals = {
+
+class Texture(Delegate):
+    id: TextureID
+    name: Optional[str] = "Unnamed Texture Delegate"
+    image: ImageID
+    sampler: Optional[SamplerID] = None
+
+
+class Sampler(Delegate):
+    id: SamplerID
+    name: Optional[str] = "Unnamed Sampler Delegate"
+
+    mag_filter: Optional[Literal["NEAREST", "LINEAR"]] = "LINEAR"
+    min_filter: Optional[Literal["NEAREST", "LINEAR", "LINEAR_MIPMAP_LINEAR"]] = "LINEAR_MIPMAP_LINEAR"
+
+    wrap_s: Optional[SamplerMode] = "REPEAT" 
+    wrap_t: Optional[SamplerMode] = "REPEAT" 
+
+
+class Light(Delegate):
+    id: LightID
+    name: Optional[str] = "Unnamed Light Delegate"
+
+    color: Optional[RGB] = [1.0, 1.0, 1.0]
+    intensity: Optional[float] = 1.0
+
+    point: PointLight = None
+    spot: SpotLight = None
+    directional: DirectionalLight = None
+
+    @root_validator
+    def one_of(cls, values):
+        already_found  = False
+        for field in ['point', 'spot', 'directional']:
+            if values[field] and already_found:
+                raise ValueError("More than one field entered")
+            elif values[field]:
+                already_found = True
+        
+        if not already_found:
+            raise ValueError("No field provided")
+        else:
+            return values
+
+
+class Geometry(Delegate):
+    id: GeometryID
+    name: Optional[str] = "Unnamed Geometry Delegate"
+    patches: list[GeometryPatch]
+
+
+class Table(Delegate):
+    id: TableID
+    name: Optional[str] = "Unnamed Table Delegate"
+
+    meta: Optional[str] = None
+    methods_list: Optional[list[MethodID]] = None
+    signals_list: Optional[list[SignalID]] = None
+
+    def __init__(self, **kwargs):
+        """Overide init to set default values"""
+
+        super.__init__(**kwargs)
+        self.signals: dict[str, Callable] = {
             "tbl_reset" : self._reset_table,
             "tbl_rows_removed" : self._remove_rows,
             "tbl_updated" : self._update_rows,
             "tbl_selection_updated" : self._update_selection
         }
-        # Specify public methods 
-        self.__all__ = [
-            "subscribe", 
-            "request_clear", 
-            "request_insert", 
-            "request_remove", 
-            "request_update", 
-            "request_update_selection",
-            "plot"
-        ]
+        self.__all__: list[str] = [
+                "subscribe", 
+                "request_clear", 
+                "request_insert", 
+                "request_remove", 
+                "request_update", 
+                "request_update_selection",
+                "plot"
+            ]
 
-
-    def _on_table_init(self, init_info: Message, on_done=None):
+    def _on_table_init(self, init_info: dict, on_done=None):
         """Creates table from server response info
 
         Args:
@@ -306,8 +567,8 @@ class TableDelegate(Delegate):
         """
 
         # Extract data from init info and transpose rows to cols
-        row_data = getattr(init_info, "data")
-        cols = getattr(init_info, "columns")
+        row_data = init_info["data"]
+        cols = init_info["columns"]
         print(f"Table Initialized with cols: {cols} and row data: {row_data}")
 
 
@@ -363,7 +624,7 @@ class TableDelegate(Delegate):
         print(f"Made selection {selection_obj.name} = {selection_obj}")
 
 
-    def _relink_signals(self):
+    def relink_signals(self):
         """Relink the signals for built in methods
 
         These should always be linked, along with whatever is injected,
@@ -376,7 +637,7 @@ class TableDelegate(Delegate):
         self.signals["noo::tbl_selection_updated"] = self._update_selection
 
 
-    def on_new(self, message: Message):
+    def on_new(self, message: dict):
         """Handler when create message is received
 
         Args:
@@ -384,10 +645,8 @@ class TableDelegate(Delegate):
         """
         
         # Set name
-        name = message["name"]
-        methods = message["methods_list"]
-        signals = message["signals_list"]
-        if name: self.name = name
+        methods = self.methods_list
+        signals = self.signals_list
     
         # Inject methods and signals
         if methods: inject_methods(self, methods)
@@ -395,20 +654,20 @@ class TableDelegate(Delegate):
 
         # Reset
         self._reset_table()
-        self._relink_signals()
+        self.relink_signals()
 
 
-    def on_update(self, message: Message):
+    def on_update(self, message: dict):
         """Handler when update message is received
         
         Args:
             message (Message): update message with the new table's info
         """
 
-        self._relink_signals()
+        self.relink_signals()
     
 
-    def on_remove(self, message: Message):
+    def on_remove(self, message: dict):
         pass
 
 
@@ -509,36 +768,154 @@ class TableDelegate(Delegate):
 
         self.tbl_update_selection(on_done, name, {"rows": keys})
 
-
-class DocumentDelegate(Delegate):
-    pass
     
-class EntityDelegate(Delegate):
-    pass
+    
+ 
 
-class PlotDelegate(Delegate):
-    pass
+""" ====================== Communication Objects ====================== """
 
-class MaterialDelegate(Delegate):
-    pass
 
-class GeometryDelegate(Delegate):
-    pass
+class Invoke(NoodleObject):
+    id: SignalID
+    context: Optional[InvokeIDType] = None # if empty - document
+    signal_data: list[Any]
 
-class LightDelegate(Delegate):
-    pass
 
-class ImageDelegate(Delegate):
-    pass
+# Note: this isn't technically an exception
+# for now this uses a model so that it can be validated / sent as message easier
+class MethodException(NoodleObject):
+    code: int
+    message: Optional[str] = None
+    data: Optional[Any] = None
 
-class TextureDelegate(Delegate):
-    pass
+class Reply(NoodleObject):
+    invoke_id: str
+    result: Optional[Any] = None
+    method_exception: Optional[MethodException] = None
 
-class SamplerDelegate(Delegate):
-    pass
 
-class BufferDelegate(Delegate):
-    pass
+""" ====================== Miscellaneous Objects ====================== """
 
-class BufferViewDelegate(Delegate):
-    pass
+default_delegates = {
+    "entities" : Entity,
+    "tables" : Table,
+    "plots" : Plot,
+    "signals" : Signal,
+    "methods" : Method,
+    "materials" : Material,
+    "geometries" : Geometry,
+    "lights" : Light,
+    "images" : Image,
+    "textures" : Texture,
+    "samplers" : Sampler,
+    "buffers" : Buffer,
+    "bufferviews" : BufferView,
+    "document" : Document
+}
+
+
+id_map = {
+    Method: MethodID,
+    Signal: SignalID,
+    Table: TableID,
+    Plot: PlotID,
+    Entity: EntityID,
+    Material: MaterialID,
+    Geometry: GeometryID,
+    Light: LightID,
+    Image: ImageID,
+    Texture: TextureID,
+    Sampler: SamplerID,
+    Buffer: BufferID,
+    BufferView: BufferViewID,
+    Document: None
+}
+
+class InjectedMethod(object):
+    """Class for representing injected method in delegate
+
+    Attributes:
+        method (method): method to be called
+        injected (bool): attribute marking method as injected
+    """
+
+    def __init__(self, method_obj) -> None:
+        self.method = method_obj
+        self.injected = True
+
+    def __call__(self, *args, **kwds):
+        self.method(*args, **kwds)
+
+
+class LinkedMethod(object):
+    """Class linking target delegate and method's delegate 
+        
+    make a cleaner function call in injected method
+    
+    Attributes:
+        _obj_delegate (delegate): 
+            delgate method is being linked to
+        _method_delegate (MethodDelegate): 
+            the method's delegate 
+    """
+
+    def __init__(self, object_delegate: Delegate, method_delegate: Delegate):
+        self._obj_delegate = object_delegate
+        self._method_delegate = method_delegate
+
+    def __call__(self, on_done=None, *arguments):
+        self._method_delegate.invoke(self._obj_delegate, arguments, callback=on_done)
+
+
+def inject_methods(delegate: Delegate, methods: list[MethodID]):
+    """Inject methods into a delegate class
+
+    Args:
+        delegate_name (str): 
+            identifier for delegate to be modified
+        methods (list): 
+            list of method id's to inject
+    """
+
+    # Clear out old injected methods
+    for name in dir(delegate):
+        att = getattr(delegate, name)
+        if hasattr(att, "injected"):
+            print(f"Deleting: {name} in inject methods")
+            delattr(delegate, name)
+
+    for id in methods:
+
+        # Get method delegate and manipulate name to exclude noo::
+        method = delegate.client.state[id]
+        name = method.name[5:]
+
+        # Create injected by linking delegates, and creating call method
+        linked = LinkedMethod(delegate, method)
+        injected = InjectedMethod(linked.__call__)
+
+        setattr(delegate, name, injected)
+
+
+def inject_signals(delegate: Delegate, signals: list[SignalID]):
+    """Method to inject signals into delegate
+
+    Args:
+        delegate (delegate): 
+            delegate object to be injected 
+        signals (list): 
+            list of signal id's to be injected
+    """
+
+    # state_signals = delegate.client.state["signals"]
+    # injected_signals = {}
+    # for id in signals:
+    #     signal: Signal = state_signals[tuple(id)]
+    #     injected_signals[signal.name] = None # Check seems strange...
+    # delegate.signals = injected_signals
+
+    for signal_id in signals:
+        signal = delegate.client.state[signal_id] # refactored state
+        delegate.signals[signal.name] = None
+
+
