+++ penne/handlers.py
@@ -1,19 +1,17 @@
 """Module for Handling Raw Messages from the Server"""
 
 from __future__ import annotations
-import asyncio
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 if TYPE_CHECKING:
-    from penne.delegates import Delegate
-    from penne.messages import Message
     from penne.core import Client
 
 import weakref
 
-from . import messages
+from penne.delegates import Delegate, id_map, default_delegates
+from penne.delegates import TableID, PlotID, EntityID
 
 # Helper Methods
-def handle_update(client, message: Message, specifier: str):
+def handle_update(client, message: dict, specifier: str):
     """Update a delegate in the current state
 
     Args:
@@ -30,7 +28,7 @@ def handle_update(client, message: Message, specifier: str):
         setattr(current_state, attribute, value)
 
 
-def delegate_from_context(client: Client, context: Message) -> Delegate:
+def delegate_from_context(client: Client, context: dict) -> Delegate:
     """Get delegate object from a context message object
     
     Args:
@@ -43,19 +41,25 @@ def delegate_from_context(client: Client, context: Message) -> Delegate:
 
     if not context:
         target_delegate = client.state["document"]
-    elif hasattr(context, "table"):
-        target_delegate = client.state["tables"][context.table]
+        return target_delegate
+
+    table = context.get("table")
+    entity = context.get("entity")
+    plot = context.get("plot")
+
+    if table:
+        target_delegate = client.state[TableID(table)]
     elif hasattr(context, "entity"):
-        target_delegate = client.state["entities"][context.entity]
+        target_delegate = client.state[EntityID(entity)]
     elif hasattr(context, "plot"):
-        target_delegate = client.state["plots"][context.plot]
+        target_delegate = client.state[PlotID(plot)]
     else:
         raise Exception("Couldn't get delegate from context")
     
     return target_delegate
 
 
-def handle(client: Client, id, message_dict):
+def handle(client: Client, message_id, message: dict[str, Any]):
     """Handle message from server
 
     'Handle' uses the ID attached to message to get handling info, and uses this info 
@@ -73,78 +77,79 @@ def handle(client: Client, id, message_dict):
     """
     
     # Process message using ID from dict
-    handle_info = client.server_message_map[id]
+    handle_info = client.server_messages[message_id]
     action = handle_info.action
     specifier = handle_info.specifier
-    message_obj: Message = messages.Message.from_dict(message_dict)
+    id_type = id_map[default_delegates[specifier]]
     print(f"Message: {action} {specifier}")
-
-    if specifier == "plots":
-        print(f"\n  {action} - {specifier}\n{message_obj}")
     
     # Update state based on map info
     if action == "create":
 
         # Create instance of delegate
-        specifier = specifier
         reference = weakref.ref(client)
         reference_obj = reference()
-        delegate: Delegate = client.delegates[specifier](reference_obj, message_obj, specifier)
+        delegate: Delegate = client.delegates[specifier](client=reference_obj, **message)
 
         # Update state and pass message info to the delegate's handler
-        client.state[specifier][message_obj.id] = delegate
-        delegate.on_new(message_obj)
+        client.state[delegate.id] = delegate
+        delegate.on_new(message)
     
     elif action == "delete":
-
-        state_delegate: Delegate = client.state[specifier][message_obj.id]
+        
+        id = id_type(message["id"])
+        state_delegate: Delegate = client.state[id]
 
         # Update delegate and state
-        state_delegate.on_remove(message_obj)
+        state_delegate.on_remove(message)
         del state_delegate
 
     elif action == "update":
 
         if specifier != "document":
-            handle_update(client, message_obj, specifier)
-            client.state[specifier][message_obj.id].on_update(message_obj)
+            id = id_type(message["id"])
+            handle_update(client, message, specifier)
+            client.state[id].on_update(message)
         else:
-            client.state[specifier].on_update(message_obj)
+            client.state[specifier].on_update(message)
 
     elif action == "reply":
 
         # Handle callback functions
-        if hasattr(message_obj, "method_exception"):
-            raise Exception(f"Method call ({message_obj.invoke_id}) resulted in exception from server: {message_obj.method_exception}")
+        exception = message.get("method_exception", False)
+        invoke_id = message.get("invoke_id")
+        result = message.get("result")
+
+        if exception:
+            raise Exception(f"Method call ({invoke_id}) resulted in exception from server: {exception}")
         else:
-            callback = client.callback_map.pop(message_obj.invoke_id)
+            callback = client.callback_map.pop(invoke_id)
             if callback:
             
-                callback_info = (callback, message_obj.result) if hasattr(message_obj, "result") else (callback, None)
+                callback_info = (callback, result)
                 client.callback_queue.put(callback_info)
-                #callback(message_obj.result) if hasattr(message_obj, "result") else callback()
 
 
     elif action == "invoke":
 
         # Handle invoke message from server
-        signal_data = message_obj.signal_data
-        signal: Delegate = client.state["signals"][message_obj.id]
+        signal_data = message["signal_data"]
+        id = id_type(message["id"])
+        signal: Delegate = client.state[id]
 
         # Determine the delegate the signal is being invoked on
-        context = getattr(message_obj, "context", False)
+        context = message.get("context")
         target_delegate = delegate_from_context(client, context)
 
         # Invoke signal attached to target delegate
-        print(f"Invoking {signal.info.name} w/ args: {signal_data}")
-        target_delegate.signals[signal.info.name](*signal_data)
+        print(f"Invoking {signal.name} w/ args: {signal_data}")
+        target_delegate.signals[signal.name](*signal_data)
 
     elif action == "initialized":
 
         if client.on_connected:
             client.callback_queue.put((client.on_connected, None))
-            #client.on_connected()
 
     else:
         # Document reset messages
-        print(message_obj)
+        print(message)
